I have added many images to this file set.
I have plans of designing websites for my portfolio.  One is centered on my family tree.  I have many files my father left me upon his death that have hand-written notes about all sorts of family members going back to the beginning of America.  I want to pull them together in an easy format for anyone in the family to see.  I need to find the time for this.  As I worked through the options in Bootstrap during this homework assignment I started getting ideas for this website.  I'd like to see more creativity in the homework.  Although that takes time and the homework is already taking a lot of time.